implements Todo instances in displayModule
The display module was using texts as todos because it was more simple to think about the structure with that. I've now implemented the class instance that i created at the beggining. This will allow to implement other features like editing todos more easily, and bring the project closer to the final idea.