moves todo addition into another module
the display module was using too much logic, so i moved the validation and addition / edition of todos to a new module: the todo interface. I'll try to use this module as a mediator between the controller and the display, with some logic that doesnt fit in neither of the previously mentioned.