implements concept of dynamic todo
the dynamic todo is the one added to a category dynamically, which means, added because its due date is coming in the next two weeks for the case of the upcoming category, or that the todo has a high priority in the case of the important category (this last one will be implemented soon).
when a dynamic todo doesn't meet the requirement to be in the category anymore (for example, an upcoming todo that has expired), gets removed from the category.